split models.py file
admin -> LocalGroup, Engineering
campaign -> Campaign, Technical_campaign
client -> Client, ClientFile
project -> Project, Technical_details
misc -> Mailing
