+++ somsolet/models/__init__.py
@@ -1,0 +1,5 @@
+from .admin import LocalGroup, Engineering
+from .campaign import Campaign, Technical_campaign
+from .client import Client, ClientFile
+from .project import Project, Technical_details
+from .misc import Mailing
