+++ somsolet/models/choices_options.py
@@ -1,4 +1,6 @@
 from django.utils.translation import gettext_lazy as _
+
+
 ITEM_STATUS = (
               ('empty status', _('Initial')),
               ('preregistered', _('Pre-Registered')),
