+++ somsolet/models/project.py
@@ -2,342 +2,12 @@ from django.contrib.auth.models import User
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
+from .campaign import Campaign
+from .client import Client
 from .choices_options import (BATERY_BRAND, INVERSOR_BRAND, ITEM_ANGLES,
                               ITEM_COMMUNITY, ITEM_DISCARDED_TYPES,
                               ITEM_ORIENTATION, ITEM_STATUS, ITEM_WARNINGS,
                               LANGUAGES, PANELS_BRAND, PANELS_TYPE)
-
-
-class LocalGroup(models.Model):
-    name = models.CharField(
-        blank=True,
-        max_length=100,
-        verbose_name=_('Name'),
-    )
-
-    phone_number = models.CharField(
-        blank=True,
-        max_length=9,
-        verbose_name=_('Telephone number'),
-    )
-
-    email = models.CharField(
-        blank=True,
-        max_length=50,
-        verbose_name=_('Email'),
-    )
-
-    language = models.CharField(
-        default='ca',
-        choices=LANGUAGES,
-        max_length=5,
-        verbose_name=_('Language'),
-    )
-
-    def __str__(self):
-        return self.name
-
-
-class Engineering(models.Model):
-    user = models.OneToOneField(
-        User,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        verbose_name=_('User'),
-    )
-
-    name = models.CharField(
-        blank=True,
-        max_length=50,
-        verbose_name=_('Name'),
-    )
-
-    tin = models.CharField(
-        blank=True,
-        max_length=9,
-        verbose_name=_('Tax identification number'),
-    )
-
-    address = models.CharField(
-        blank=True,
-        max_length=100,
-        verbose_name=_('Address'),
-    )
-
-    email = models.CharField(
-        blank=True,
-        max_length=50,
-        verbose_name=_('Email'),
-    )
-
-    phone_number = models.CharField(
-        blank=True,
-        max_length=9,
-        verbose_name=_('Telephone number'),
-    )
-
-    count_closed_campaigns = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Closed Campaigns'),
-    )
-
-    count_open_campaigns = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Open Campaigns'),
-    )
-
-    count_closed_projects = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Closed Installations'),
-    )
-
-    total_kwp_installed = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Total kWp installed'),
-    )
-
-    comments = models.CharField(
-        blank=True,
-        max_length=500,
-        verbose_name=_('Comments'),
-    )
-
-    language = models.CharField(
-        default='ca',
-        choices=LANGUAGES,
-        max_length=5,
-        verbose_name=_('Language'),
-    )
-
-    def __str__(self):
-        return self.name
-
-
-class Campaign(models.Model):
-    name = models.CharField(
-        blank=True,
-        max_length=50,
-        verbose_name=_('Name'),
-    )
-
-    engineerings = models.ManyToManyField(
-        Engineering,
-        related_name='campaigns',
-        verbose_name=_('Engineering'))
-
-    local_group = models.ManyToManyField(
-        LocalGroup,
-        related_name='campaigns',
-        verbose_name=_('LocalGroup'))
-
-    date_call_for_engineerings = models.DateField(
-        null=True,
-        blank=True,
-        verbose_name=_('Date call for engineerings'),
-    )
-
-    date_call_for_inscriptions = models.DateField(
-        null=True,
-        blank=True,
-        verbose_name=_('date call for inscriptions'),
-    )
-
-    date_inscriptions_closed = models.DateField(
-        null=True,
-        blank=True,
-        verbose_name=_('Date inscriptions closed'),
-    )
-
-    date_completed_installations = models.DateField(
-        null=True,
-        blank=True,
-        verbose_name=_('Date completed installations'),
-    )
-
-    autonomous_community = models.CharField(
-        choices=ITEM_COMMUNITY,
-        default='empty',
-        max_length=50,
-        verbose_name=_('Autonomous community'),
-    )
-
-    geographical_region = models.CharField(
-        max_length=50,
-        blank=True,
-        verbose_name=_('Geographical region'),
-    )
-
-    count_foreseen_installations = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Number of planned installations'),
-    )
-
-    count_completed_installations = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Count completed installations'),
-    )
-
-    kwp_installed = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('kWp installed'),
-    )
-
-    date_warranty_payment = models.DateField(
-        null=True,
-        blank=True,
-        verbose_name=_('Date warranty payment'),
-    )
-
-    warranty_pending_amount = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Pending Warranty (€)'),
-    )
-
-    warranty_payed_amount = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Payed Warranty (€)'),
-    )
-
-    total_penalties_days = models.IntegerField(
-        null=True,
-        blank=True,
-        verbose_name=_('Total Penalties (days)'),
-    )
-
-    total_amount_penalties = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Total Penalties (€)'),
-    )
-
-    active = models.BooleanField(
-        default=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Technical_campaign(models.Model):
-    campaign = models.ForeignKey(
-        Campaign,
-        on_delete=models.CASCADE,
-        verbose_name=_('Campaign'),
-    )
-
-    price_mono_fixed = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Monofasic Fixed Price (€)'),
-    )
-
-    price_mono_var = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Monofasic Variable Price (€/Wp)'),
-    )
-
-    price_tri_fixed = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Trifasic Fixed Price (€)'),
-    )
-
-    price_tri_var = models.FloatField(
-        null=True,
-        blank=True,
-        verbose_name=_('Trifasic Variable Price (€/Wp)'),
-    )
-
-    def __str__(self):
-        return self.campaign.name
-
-
-class ClientFile(models.Model):
-    name = models.CharField(
-        blank=True,
-        max_length=100,
-        verbose_name=_('Filename'),
-    )
-
-    file = models.FileField(
-        upload_to='uploaded_files_som/general_conditions',
-        null=True,
-        blank=True,
-        verbose_name=_('Upload general conditions')
-    )
-
-    language = models.CharField(
-        default='ca',
-        choices=LANGUAGES,
-        max_length=5,
-        verbose_name=_('Language')
-    )
-
-    def __str__(self):
-        return self.name
-
-
-class Client(models.Model):
-    name = models.CharField(
-        blank=True,
-        max_length=100,
-        verbose_name=_('Name'),
-    )
-
-    membership_number = models.CharField(
-        blank=True,
-        max_length=100,
-        verbose_name=_('Membership number'),
-    )
-
-    dni = models.CharField(
-        blank=True,
-        max_length=9,
-        verbose_name=_('DNI'),
-    )
-
-    phone_number = models.CharField(
-        blank=True,
-        max_length=9,
-        verbose_name=_('Telephone number'),
-    )
-
-    email = models.CharField(
-        blank=True,
-        max_length=50,
-        verbose_name=_('Email'),
-    )
-
-    language = models.CharField(
-        default='ca',
-        choices=LANGUAGES,
-        max_length=5,
-        verbose_name=_('Language'),
-    )
-
-    sent_general_conditions = models.BooleanField(
-        default=False,
-        verbose_name=_('General conditions sent')
-    )
-
-    file = models.ManyToManyField(
-        ClientFile,
-        related_name='clients',
-        verbose_name=_('File')
-    )
-
-    def __str__(self):
-        return self.name
 
 
 class Project(models.Model):
@@ -878,26 +548,3 @@ class Technical_details(models.Model):
 
     def __str__(self):
         return self.project.name
-
-
-class Mailing(models.Model):
-
-    project = models.ForeignKey(
-        Project,
-        null=True,
-        blank=True,
-        on_delete=models.CASCADE,
-        verbose_name=_('Project')
-    )
-
-    notification_status = models.CharField(
-        choices=ITEM_STATUS,
-        default=_('empty status'),
-        max_length=50
-    )
-
-    sent = models.BooleanField(
-        default=False,
-        verbose_name=_('Sent'),
-        help_text=_('True if an email was sent')
-    )
